Synchronization of GetHistory results with multiple symbols in Live mode
Previously, in Live mode, requesting History for multiple symbols in a single call returned an IEnumerable<Slice> concatenating the results of each request, so every Slice only had one data point and the list was not sorted by time.

This problem has been solved in the following steps:
- Added a new BrokerageHistoryProvider, to be used in live mode with all brokerages
- Removed IHistoryProvider implementations from IB, Oanda, FXCM brokerages
- Added new GetHistory(request) method to IBrokerage interface and implementations
- Moved CreateSliceEnumerableFromSubscriptions method from SubscriptionDataReaderHistoryProvider to the new abstract base class SynchronizingHistoryProvider, to enable reuse in BrokerageHistoryProvider
