+++ Common/Interfaces/IBrokerage.cs
@@ -16,6 +16,8 @@
 using System;
 using System.Collections.Generic;
 using QuantConnect.Brokerages;
+using QuantConnect.Data;
+using QuantConnect.Data.Market;
 using QuantConnect.Orders;
 using QuantConnect.Securities;
 
@@ -110,5 +112,12 @@ namespace QuantConnect.Interfaces
         /// Specifies whether the brokerage will instantly update account balances
         /// </summary>
         bool AccountInstantlyUpdated { get; }
+
+        /// <summary>
+        /// Gets the history for the requested security
+        /// </summary>
+        /// <param name="request">The historical data request</param>
+        /// <returns>An enumerable of bars covering the span specified in the request</returns>
+        IEnumerable<BaseData> GetHistory(HistoryRequest request);
     }
 }
