+++ Engine/HistoricalData/SubscriptionDataReaderHistoryProvider.cs
@@ -18,11 +18,9 @@ using System;
 using System.Collections;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Threading;
 using NodaTime;
 using QuantConnect.Data;
 using QuantConnect.Data.Auxiliary;
-using QuantConnect.Data.UniverseSelection;
 using QuantConnect.Interfaces;
 using QuantConnect.Lean.Engine.DataFeeds;
 using QuantConnect.Lean.Engine.DataFeeds.Enumerators;
@@ -42,21 +40,12 @@ namespace QuantConnect.Lean.Engine.HistoricalData
     /// Provides an implementation of <see cref="IHistoryProvider"/> that uses <see cref="BaseData"/>
     /// instances to retrieve historical data
     /// </summary>
-    public class SubscriptionDataReaderHistoryProvider : IHistoryProvider
+    public class SubscriptionDataReaderHistoryProvider : SynchronizingHistoryProvider
     {
-        private int _dataPointCount;
         private IMapFileProvider _mapFileProvider;
         private IFactorFileProvider _factorFileProvider;
         private IDataProvider _dataProvider;
         private IDataCacheProvider _dataCacheProvider;
-
-        /// <summary>
-        /// Gets the total number of data points emitted by this history provider
-        /// </summary>
-        public int DataPointCount
-        {
-            get { return _dataPointCount; }
-        }
 
         /// <summary>
         /// Initializes this history provider to work for the specified job
@@ -67,7 +56,7 @@ namespace QuantConnect.Lean.Engine.HistoricalData
         /// <param name="dataProvider">Provider used to get data when it is not present on disk</param>
         /// <param name="statusUpdate">Function used to send status updates</param>
         /// <param name="dataCacheProvider">Provider used to cache history data files</param>
-        public void Initialize(AlgorithmNodePacket job, IDataProvider dataProvider, IDataCacheProvider dataCacheProvider, IMapFileProvider mapFileProvider, IFactorFileProvider factorFileProvider, Action<int> statusUpdate)
+        public override void Initialize(AlgorithmNodePacket job, IDataProvider dataProvider, IDataCacheProvider dataCacheProvider, IMapFileProvider mapFileProvider, IFactorFileProvider factorFileProvider, Action<int> statusUpdate)
         {
             _mapFileProvider = mapFileProvider;
             _factorFileProvider = factorFileProvider;
@@ -81,7 +70,7 @@ namespace QuantConnect.Lean.Engine.HistoricalData
         /// <param name="requests">The historical data requests</param>
         /// <param name="sliceTimeZone">The time zone used when time stamping the slice instances</param>
         /// <returns>An enumerable of the slices of data covering the span specified in each request</returns>
-        public IEnumerable<Slice> GetHistory(IEnumerable<HistoryRequest> requests, DateTimeZone sliceTimeZone)
+        public override IEnumerable<Slice> GetHistory(IEnumerable<HistoryRequest> requests, DateTimeZone sliceTimeZone)
         {
             // create subscription objects from the configs
             var subscriptions = new List<Subscription>();
@@ -158,70 +147,6 @@ namespace QuantConnect.Lean.Engine.HistoricalData
 
             var timeZoneOffsetProvider = new TimeZoneOffsetProvider(security.Exchange.TimeZone, start, end);
             return new Subscription(null, security, config, reader, timeZoneOffsetProvider, start, end, false);
-        }
-
-        /// <summary>
-        /// Enumerates the subscriptions into slices
-        /// </summary>
-        private IEnumerable<Slice> CreateSliceEnumerableFromSubscriptions(List<Subscription> subscriptions, DateTimeZone sliceTimeZone)
-        {
-            // required by TimeSlice.Create, but we don't need it's behavior
-            var cashBook = new CashBook();
-            cashBook.Clear();
-            var frontier = DateTime.MinValue;
-            while (true)
-            {
-                var earlyBirdTicks = long.MaxValue;
-                var data = new List<DataFeedPacket>();
-                foreach (var subscription in subscriptions)
-                {
-                    if (subscription.EndOfStream) continue;
-
-                    var packet = new DataFeedPacket(subscription.Security, subscription.Configuration);
-
-                    var offsetProvider = subscription.OffsetProvider;
-                    var currentOffsetTicks = offsetProvider.GetOffsetTicks(frontier);
-                    while (subscription.Current.EndTime.Ticks - currentOffsetTicks <= frontier.Ticks)
-                    {
-                        // we want bars rounded using their subscription times, we make a clone
-                        // so we don't interfere with the enumerator's internal logic
-                        var clone = subscription.Current.Clone(subscription.Current.IsFillForward);
-                        clone.Time = clone.Time.RoundDown(subscription.Configuration.Increment);
-                        packet.Add(clone);
-                        Interlocked.Increment(ref _dataPointCount);
-                        if (!subscription.MoveNext())
-                        {
-                            break;
-                        }
-                    }
-                    // only add if we have data
-                    if (packet.Count != 0) data.Add(packet);
-                    // udate our early bird ticks (next frontier time)
-                    if (subscription.Current != null)
-                    {
-                        // take the earliest between the next piece of data or the next tz discontinuity
-                        var nextDataOrDiscontinuity = Math.Min(subscription.Current.EndTime.Ticks - currentOffsetTicks, offsetProvider.GetNextDiscontinuity());
-                        earlyBirdTicks = Math.Min(earlyBirdTicks, nextDataOrDiscontinuity);
-                    }
-                }
-
-                // end of subscriptions
-                if (earlyBirdTicks == long.MaxValue) break;
-
-                if (data.Count != 0)
-                {
-                    // reuse the slice construction code from TimeSlice.Create
-                    yield return TimeSlice.Create(frontier, sliceTimeZone, cashBook, data, SecurityChanges.None).Slice;
-                }
-
-                frontier = new DateTime(Math.Max(earlyBirdTicks, frontier.Ticks), DateTimeKind.Utc);
-            }
-
-            // make sure we clean up after ourselves
-            foreach (var subscription in subscriptions)
-            {
-                subscription.Dispose();
-            }
         }
 
         // this implementation is provided solely for the data reader's dependency,
