+++ Tests/Brokerages/Fxcm/FxcmBrokerageHistoryProviderTests.cs
@@ -17,8 +17,10 @@ using System;
 using System.Diagnostics;
 using System.Linq;
 using NUnit.Framework;
+using QuantConnect.Brokerages;
 using QuantConnect.Brokerages.Fxcm;
 using QuantConnect.Data;
+using QuantConnect.Lean.Engine.HistoricalData;
 using QuantConnect.Logging;
 
 namespace QuantConnect.Tests.Brokerages.Fxcm
@@ -57,6 +59,11 @@ namespace QuantConnect.Tests.Brokerages.Fxcm
             TestDelegate test = () =>
             {
                 var brokerage = (FxcmBrokerage) Brokerage;
+
+                var historyProvider = new BrokerageHistoryProvider();
+                historyProvider.SetBrokerage(brokerage);
+                historyProvider.Initialize(null, null, null, null, null, null);
+
                 var now = DateTime.UtcNow;
 
                 var requests = new[]
@@ -70,7 +77,7 @@ namespace QuantConnect.Tests.Brokerages.Fxcm
                     }
                 };
 
-                var history = brokerage.GetHistory(requests, TimeZones.Utc);
+                var history = historyProvider.GetHistory(requests, TimeZones.Utc);
 
                 foreach (var slice in history)
                 {
@@ -89,7 +96,7 @@ namespace QuantConnect.Tests.Brokerages.Fxcm
                     }
                 }
 
-                Log.Trace("Data points retrieved: " + brokerage.DataPointCount);
+                Log.Trace("Data points retrieved: " + historyProvider.DataPointCount);
             };
 
             if (throwsException)
@@ -113,6 +120,10 @@ namespace QuantConnect.Tests.Brokerages.Fxcm
 
             var brokerage = (FxcmBrokerage)Brokerage;
 
+            var historyProvider = new BrokerageHistoryProvider();
+            historyProvider.SetBrokerage(brokerage);
+            historyProvider.Initialize(null, null, null, null, null, null);
+
             var stopwatch = Stopwatch.StartNew();
 
             var requests = new[]
@@ -126,16 +137,16 @@ namespace QuantConnect.Tests.Brokerages.Fxcm
                 }
             };
 
-            var history = brokerage.GetHistory(requests, TimeZones.Utc);
+            var history = historyProvider.GetHistory(requests, TimeZones.Utc);
             var tickCount = history.SelectMany(slice => slice.Ticks[symbol]).Count();
 
             stopwatch.Stop();
             Log.Trace("Download time: " + stopwatch.Elapsed);
 
             Log.Trace("History ticks returned: " + tickCount);
-            Log.Trace("Data points retrieved: " + brokerage.DataPointCount);
+            Log.Trace("Data points retrieved: " + historyProvider.DataPointCount);
 
-            Assert.AreEqual(tickCount, brokerage.DataPointCount);
+            Assert.AreEqual(tickCount, historyProvider.DataPointCount);
             Assert.AreEqual(241233, tickCount);
         }
     }
